Змінтв розшироення,додав теми та медіазапити
