+++ src/components/Account/UserForm/readme.md
@@ -1,0 +1,17 @@
+1. Компонент рендериться на маршрут /account
+2. На сторінці знаходиться компонент форми UserForm"
+
+3. Компонент сторінки підписаний на глобальний стейт state.auth.user.
+4. Компонент рендерить форму з полями, заповненми даними про юзера та має
+   локальний стейт для цих данних.
+5. За відсутності данних, в полях форми показуються плейсхолдери, а в полі з
+   датою показується поточна дата.
+6. При додаванні картинки в поле для вибору аватара, вона з'являється на місці
+   плейсхолдера.
+7. Кнопка сабміту не активна, поки в полях не зміняться значення.
+8. Після зміни будь-якого значення кнопка стає активною.
+9. Додати валідацію на поля форми.
+10. По сабміту дані відправляються на бек і кнопка сабміту стає знову
+    неактивною.
+11. При успішній відповіді данні пишуться в глобальний стейт.
+12. При помилці юзеру виводиться відповідне пуш-повідомлення."
